updated router-view to be sfc
