+++ src/compositions/useRouterDepth.ts
@@ -1,5 +1,5 @@
 import { inject } from 'vue'
-import { depthInjectionKey } from '@/components/routerView'
+import { depthInjectionKey } from '@/types'
 
 export function useRouterDepth(): number {
   return inject(depthInjectionKey, 0)
