+++ src/types/index.ts
@@ -1,5 +1,6 @@
 export * from './errors'
 export * from './hooks'
+export * from './injectionDepth'
 export * from './invalidRouteParamValueError'
 export * from './params'
 export * from './register'
@@ -9,8 +10,8 @@ export * from './routeMatchRule'
 export * from './routeMethod'
 export * from './routeMethods'
 export * from './routePaths'
+export * from './routeWithParams'
 export * from './router'
 export * from './routerRoute'
 export * from './routes'
-export * from './routeWithParams'
 export * from './utilities'
