+++ src/types/injectionDepth.ts
@@ -1,0 +1,3 @@
+import { InjectionKey } from 'vue'
+
+export const depthInjectionKey: InjectionKey<number> = Symbol()
